requirements.txt
asgiref==3.4.1
astropy==4.3.1
astroquery==0.4.3
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
cycler==0.10.0
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==1.16.0
gunicorn==20.1.0
html5lib==1.1
idna==3.2
importlib-metadata==4.8.1
jplephem==2.16
keyring==23.2.1
kiwisolver==1.3.2
llvmlite==0.37.0
matplotlib==3.4.3
mimeparse==0.1.3
numba==0.54.0
numpy==1.20.3
pandas==1.3.3
Pillow==8.3.2
plotly==5.3.1
poliastro==0.15.2
pyerfa==2.0.0
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-snappy==0.6.0
pytz==2021.1
pyvo==1.1
pywin32-ctypes==0.2.0
requests==2.26.0
scipy==1.7.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.2.4
tenacity==8.0.1
urllib3==1.26.6
webencodings==0.5.1
zipp==3.5.1
